Endpoint subcategorias por producto y backend conectado

+++ BackEnd/FJSN_ZGE_ABG_Proyecto_TFC_01/src/main/java/tiendaonline/repository/FavoritosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tiendaonline.entities.Favoritos;
 
-public interface FavoritosRepository extends JpaRepository<Favoritos, Long>{
+public interface FavoritosRepository extends JpaRepository<Favoritos, Integer>{
 
 }
