+++ BackEnd/FJSN_ZGE_ABG_Proyecto_TFC_01/src/main/java/tiendaonline/service/PedidoService.java
@@ -9,11 +9,11 @@ public interface PedidoService {
 	
 	List<Pedido> findAll();
 
-	Pedido findById(Long id);
+	Pedido findById(Integer id);
 
 	Pedido insertOne(Pedido pedido);
 
 	Pedido updateOne(Pedido pedido);
 
-	int deleteOne(Long id);
+	int deleteOne(Integer id);
 }
